+++ packages/weapp-tailwindcss/src/types/user-defined-options/index.ts
@@ -5,10 +5,24 @@ import type { UserDefinedOptionsMatcherPart } from './matcher'
 
 export type {
   UserDefinedOptionsGeneralPart,
+} from './general'
+
+export type {
+  UniAppXComponentLocalStylesOptions,
+  UniAppXOptions,
+} from './important'
+
+export type {
   UserDefinedOptionsImportantPart,
+} from './important'
+
+export type {
   UserDefinedOptionsLifecyclePart,
+} from './lifecycle'
+
+export type {
   UserDefinedOptionsMatcherPart,
-}
+} from './matcher'
 
 export interface UserDefinedOptions
   extends UserDefinedOptionsGeneralPart,